Interrupciones hardware y controlador PIC 8259

+++ include/cpu/regs.h
@@ -31,7 +31,7 @@
 
 struct regs
 {
-	uint32 gs, fs, es, ds;//0-12
+	uint16 gs, fs, es, ds;//0-12
 	uint32 edi, esi, ebp, esp, ebx, edx, ecx, eax;//16----32-36-40-44
 	uint32 int_no, err_code;
 	uint32 eip, cs, eflags, useresp, ss;
